+++ src/main/java/com/wenxin2/marioverse/MarioverseCreativeTabs.java
@@ -102,6 +102,8 @@ public class MarioverseCreativeTabs {
 
             add(event, ItemRegistry.MARIO_ARMOR_TRIM_SMITHING_TEMPLATE);
             add(event, ItemRegistry.LUIGI_ARMOR_TRIM_SMITHING_TEMPLATE);
+            add(event, ItemRegistry.WARIO_ARMOR_TRIM_SMITHING_TEMPLATE);
+            add(event, ItemRegistry.WALUIGI_ARMOR_TRIM_SMITHING_TEMPLATE);
             add(event, ItemRegistry.PRINCESS_ARMOR_TRIM_SMITHING_TEMPLATE);
 
             add(event, ItemRegistry.CHRISTMAS_HAT);
@@ -726,7 +728,9 @@ public class MarioverseCreativeTabs {
 
                 addAfter(event, Items.BOLT_ARMOR_TRIM_SMITHING_TEMPLATE, ItemRegistry.MARIO_ARMOR_TRIM_SMITHING_TEMPLATE);
                 addAfter(event, ItemRegistry.MARIO_ARMOR_TRIM_SMITHING_TEMPLATE, ItemRegistry.LUIGI_ARMOR_TRIM_SMITHING_TEMPLATE);
-                addAfter(event, ItemRegistry.LUIGI_ARMOR_TRIM_SMITHING_TEMPLATE, ItemRegistry.PRINCESS_ARMOR_TRIM_SMITHING_TEMPLATE);
+                addAfter(event, ItemRegistry.LUIGI_ARMOR_TRIM_SMITHING_TEMPLATE, ItemRegistry.WARIO_ARMOR_TRIM_SMITHING_TEMPLATE);
+                addAfter(event, ItemRegistry.WARIO_ARMOR_TRIM_SMITHING_TEMPLATE, ItemRegistry.WALUIGI_ARMOR_TRIM_SMITHING_TEMPLATE);
+                addAfter(event, ItemRegistry.WALUIGI_ARMOR_TRIM_SMITHING_TEMPLATE, ItemRegistry.PRINCESS_ARMOR_TRIM_SMITHING_TEMPLATE);
             }
 
             if (event.getTabKey() == CreativeModeTabs.SPAWN_EGGS) {