+++ src/main/java/com/wenxin2/marioverse/datagen/ItemModelGen.java
@@ -73,7 +73,7 @@ public class ItemModelGen extends ItemModelProvider {
         this.basicItem(ItemRegistry.ICE_FLOWER_SPAWN_EGG.get());
         this.basicItem(ItemRegistry.LARGE_SNOWBALL.get());
         this.basicItem(ItemRegistry.LUIGI_ARMOR_TRIM_SMITHING_TEMPLATE.get());
-        this.basicItem(ItemRegistry.MARIO_ARMOR_TRIM_SMITHING_TEMPLATE.get());
+        this.basicItem(ItemRegistry.MARIO_ARMOR_TRIM_SMITHING_TEMPLATE.get());;
         this.basicItem(ItemRegistry.MEGA_GOOMBA_SPAWN_EGG.get());
         this.basicItem(ItemRegistry.MINI_GOOMBA_SPAWN_EGG.get());
         this.basicItem(ItemRegistry.MINI_MUSHROOM.get());
@@ -104,6 +104,8 @@ public class ItemModelGen extends ItemModelProvider {
         this.basicItem(ItemRegistry.SUPER_MUSHROOM_SPAWN_EGG.get());
         this.basicItem(ItemRegistry.SUPER_STAR.get());
         this.basicItem(ItemRegistry.SUPER_STAR_SPAWN_EGG.get());
+        this.basicItem(ItemRegistry.WALUIGI_ARMOR_TRIM_SMITHING_TEMPLATE.get());
+        this.basicItem(ItemRegistry.WARIO_ARMOR_TRIM_SMITHING_TEMPLATE.get());
         this.basicItem(ItemRegistry.WARM_CHEEP_CHEEP.get());
         this.basicItem(ItemRegistry.WHITE_KOOPA_SHOES.get());
         this.basicItem(ItemRegistry.WHITE_KOOPA_SHOES.get());