+++ src/main/java/com/wenxin2/marioverse/datagen/RecipeGen.java
@@ -82,6 +82,8 @@ public class RecipeGen extends RecipeUtils {
         copySmithingTemplate(output, ItemRegistry.LUIGI_ARMOR_TRIM_SMITHING_TEMPLATE, Ingredient.of(Items.LIME_WOOL, Items.GREEN_WOOL));
         copySmithingTemplate(output, ItemRegistry.MARIO_ARMOR_TRIM_SMITHING_TEMPLATE, Ingredient.of(Items.RED_WOOL));
         copySmithingTemplate(output, ItemRegistry.PRINCESS_ARMOR_TRIM_SMITHING_TEMPLATE, Ingredient.of(Items.PINK_WOOL));
+        copySmithingTemplate(output, ItemRegistry.WALUIGI_ARMOR_TRIM_SMITHING_TEMPLATE, Ingredient.of(Items.PURPLE_WOOL));
+        copySmithingTemplate(output, ItemRegistry.WARIO_ARMOR_TRIM_SMITHING_TEMPLATE, Ingredient.of(Items.YELLOW_WOOL));
 
         checkeredRecipe(4, "spike_panels", BlockRegistry.CALCITE_CHECKERED_TILES, BlockRegistry.POLISHED_CALCITE.get(DyeColor.WHITE), BlockRegistry.POLISHED_CALCITE.get(DyeColor.BLACK), false, output);
         classicCheckpointFlagRecipe(1, BlockRegistry.CLASSIC_CHECKPOINT_FLAG, Tags.Items.INGOTS_GOLD, ItemRegistry.SUPER_MUSHROOM, Items.WHITE_WOOL, Blocks.SMOOTH_STONE_SLAB, output);