+++ src/main/java/com/wenxin2/marioverse/registries/TrimPatternRegistry.java
@@ -10,6 +10,8 @@ public class TrimPatternRegistry {
     public static final ResourceKey<TrimPattern> LUIGI;
     public static final ResourceKey<TrimPattern> MARIO;
     public static final ResourceKey<TrimPattern> PRINCESS;
+    public static final ResourceKey<TrimPattern> WALUIGI;
+    public static final ResourceKey<TrimPattern> WARIO;
 
     static {
         LUIGI = ResourceKey.create(Registries.TRIM_PATTERN,
@@ -18,5 +20,9 @@ public class TrimPatternRegistry {
                 ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "mario"));
         PRINCESS = ResourceKey.create(Registries.TRIM_PATTERN,
                 ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "princess"));
+        WALUIGI = ResourceKey.create(Registries.TRIM_PATTERN,
+                ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "waluigi"));
+        WARIO = ResourceKey.create(Registries.TRIM_PATTERN,
+                ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "wario"));
     }
 }